Cleaned up and made changes according to the review comments
