+++ code/go/0chain.net/smartcontract/dbs/event/reward_provider.go
@@ -60,111 +60,3 @@ func (edb *EventDb) GetProviderRewards(limit common.Pagination, id string, start
 			Desc:   limit.IsDescending,
 		}).Scan(&rps).Error
 }
-
-func (edb *EventDb) GetChallengeRewardsByChallengeID(challengeID string) []RewardProvider {
-
-	var rps []RewardProvider
-	edb.Get().Where("challenge_id = ? AND reward_type IN (6, 8, 9)", challengeID).Find(&rps)
-
-	return rps
-}
-
-func (edb *EventDb) GetChallengeRewardsByProviderID(providerID string) []RewardProvider {
-
-	var rps []RewardProvider
-	edb.Get().Where("provider_id = ? AND reward_type IN (0, 6, 8, 9, 10)", providerID).Find(&rps)
-
-	return rps
-}
-
-func (edb *EventDb) GetAllProviderChallengeRewards() []RewardProvider {
-
-	var rps []RewardProvider
-	edb.Get().Where("reward_type IN (0, 6, 8, 9, 10)").Find(&rps)
-
-	return rps
-}
-
-func (edb *EventDb) GetRewardsByRewardType(rewardType string) []RewardProvider {
-
-	var rps []RewardProvider
-	edb.Get().Where("reward_type = ?", rewardType).Find(&rps)
-
-	return rps
-}
-
-func (edb *EventDb) RunWhereQueryInProviderRewards(query string) []RewardProvider {
-
-	var rps []RewardProvider
-
-	edb.Get().Where(query).Find(&rps)
-
-	return rps
-
-}
-
-func (edb *EventDb) GetBlockRewardsToProviders(blockNumber, startBlockNumber, endBlockNumber string) []RewardProvider {
-
-	if blockNumber != "" {
-		var rps []RewardProvider
-		edb.Get().Where("block_number = ? AND reward_type IN (?)", blockNumber, spenum.BlockRewardBlobber).Find(&rps)
-
-		return rps
-	}
-
-	if startBlockNumber != "" && endBlockNumber != "" {
-		var rps []RewardProvider
-		edb.Get().Where("block_number >= ? AND block_number <= ? AND reward_type IN (?)", startBlockNumber, endBlockNumber, spenum.BlockRewardBlobber).Find(&rps)
-
-		return rps
-	}
-
-	return nil
-}
-
-func (edb *EventDb) GetReadRewardsToProviders(blockNumber, startBlockNumber, endBlockNumber string) []RewardProvider {
-
-	if blockNumber != "" {
-		var rps []RewardProvider
-		edb.Get().Where("block_number = ? AND reward_type IN (?)", blockNumber, spenum.FileDownloadReward).Find(&rps)
-
-		return rps
-	}
-
-	if startBlockNumber != "" && endBlockNumber != "" {
-		var rps []RewardProvider
-		edb.Get().Where("block_number >= ? AND block_number <= ? AND reward_type IN (?)", startBlockNumber, endBlockNumber, spenum.FileDownloadReward).Find(&rps)
-
-		return rps
-	}
-
-	return nil
-}
-
-func (edb *EventDb) GetChallengeRewardsToProviders(challengeID string) ([]RewardProvider, []RewardProvider) {
-
-	var blobberRewards []RewardProvider
-	edb.Get().Where("challenge_id = ? AND reward_type IN (?)", challengeID, spenum.ChallengePassReward).Find(&blobberRewards)
-
-	var validatorRewards []RewardProvider
-	edb.Get().Where("challenge_id = ? AND reward_type IN (?)", challengeID, spenum.ValidationReward).Find(&validatorRewards)
-
-	return blobberRewards, validatorRewards
-}
-
-func (edb *EventDb) GetTotalChallengeRewardsToProviders(challengeID string) int64 {
-
-	var totalChallengeRewards int64
-
-	edb.Get().Where("challenge_id = ? AND reward_type IN (?, ?)", challengeID, spenum.ChallengePassReward, spenum.ValidationReward).Scan(&totalChallengeRewards)
-
-	return totalChallengeRewards
-}
-
-func (edb *EventDb) GetAllocationCancellationRewardsToProviders(startBlock, endBlock string) []RewardProvider {
-
-	var rps []RewardProvider
-	edb.Get().Where("block_number >= ? AND block_number <= ? AND reward_type IN (?)", startBlock, endBlock, spenum.CancellationChargeReward).Find(&rps)
-
-	return rps
-}
