+++ code/go/0chain.net/smartcontract/storagesc/handler.go
@@ -1,6 +1,7 @@
 package storagesc
 
 import (
+	"0chain.net/chaincore/config"
 	"encoding/json"
 	"errors"
 	"fmt"
@@ -50,7 +51,7 @@ func SetupRestHandler(rh rest.RestHandlerI) {
 func GetEndpoints(rh rest.RestHandlerI) []rest.Endpoint {
 	srh := NewStorageRestHandler(rh)
 	storage := "/v1/screst/" + ADDRESS
-	return []rest.Endpoint{
+	restEndpoints := []rest.Endpoint{
 		rest.MakeEndpoint(storage+"/getBlobber", common.UserRateLimit(srh.getBlobber)),
 		rest.MakeEndpoint(storage+"/getblobbers", common.UserRateLimit(srh.getBlobbers)),
 		rest.MakeEndpoint(storage+"/blobbers-by-rank", common.UserRateLimit(srh.getBlobbersByRank)),
@@ -94,15 +95,18 @@ func GetEndpoints(rh rest.RestHandlerI) []rest.Endpoint {
 		rest.MakeEndpoint(storage+"/replicate-authorizer-aggregates", srh.replicateAuthorizerAggregates),
 		rest.MakeEndpoint(storage+"/replicate-validator-aggregates", srh.replicateValidatorAggregates),
 		rest.MakeEndpoint(storage+"/replicate-user-aggregates", srh.replicateUserAggregates),
-		rest.MakeEndpoint(storage+"/reward-providers", srh.rewardProviders),
-		rest.MakeEndpoint(storage+"/reward-delegates", srh.rewardProviders),
-		rest.MakeEndpoint(storage+"/all-challenges", srh.getAllChallenges),
-		rest.MakeEndpoint(storage+"/block-rewards", srh.getBlockRewards),
-		rest.MakeEndpoint(storage+"/read-rewards", srh.getReadRewards),
-		rest.MakeEndpoint(storage+"/challenge-rewards", srh.getChallengeRewards),
-		rest.MakeEndpoint(storage+"/total-challenge-rewards", srh.getTotalChallengeRewards),
-		rest.MakeEndpoint(storage+"/cancellation-rewards", srh.getAllocationCancellationReward),
 	}
+
+	if config.Development() {
+		restEndpoints = append(restEndpoints, rest.MakeEndpoint(storage+"/all-challenges", srh.getAllChallenges))
+		restEndpoints = append(restEndpoints, rest.MakeEndpoint(storage+"/block-rewards", srh.getBlockRewards))
+		restEndpoints = append(restEndpoints, rest.MakeEndpoint(storage+"/read-rewards", srh.getReadRewards))
+		restEndpoints = append(restEndpoints, rest.MakeEndpoint(storage+"/challenge-rewards", srh.getChallengeRewards))
+		restEndpoints = append(restEndpoints, rest.MakeEndpoint(storage+"/total-challenge-rewards", srh.getTotalChallengeRewards))
+		restEndpoints = append(restEndpoints, rest.MakeEndpoint(storage+"/cancellation-rewards", srh.getAllocationCancellationReward))
+	}
+
+	return restEndpoints
 }
 
 // swagger:route GET /v1/screst/6dba10422e368813802877a85039d3985d96760ed844092319743fb3a76712d7/blobber_ids blobber_ids
@@ -2984,257 +2988,4 @@ func (srh *StorageRestHandler) replicateUserAggregates(w http.ResponseWriter, r 
 		users = []event.UserAggregate{}
 	}
 	common.Respond(w, r, users, nil)
-}
-
-func (srh *StorageRestHandler) rewardProviders(w http.ResponseWriter, r *http.Request) {
-	// read all data from reward_providers table and return
-	edb := srh.GetQueryStateContext().GetEventDB()
-	if edb == nil {
-		common.Respond(w, r, nil, common.NewErrInternal("no db connection"))
-		return
-	}
-
-	challengeID := r.URL.Query().Get("challenge_id")
-	providerID := r.URL.Query().Get("provider_id")
-	query := r.URL.Query().Get("query")
-
-	var rps []event.RewardProvider
-
-	if query != "" {
-		rps = edb.RunWhereQueryInProviderRewards(query)
-	} else if challengeID != "" {
-		rps = edb.GetChallengeRewardsByChallengeID(challengeID)
-	} else if providerID != "" {
-		rps = edb.GetChallengeRewardsByProviderID(providerID)
-	} else {
-		rps = edb.GetAllProviderChallengeRewards()
-	}
-
-	var sum int64
-
-	for _, rp := range rps {
-		cur, _ := rp.Amount.Int64()
-		sum += cur
-	}
-
-	result := map[string]interface{}{
-		"sum": sum,
-		"rps": rps,
-	}
-
-	// return the list of reward providers
-	common.Respond(w, r, result, nil)
-}
-
-func (srh *StorageRestHandler) rewardDelegates(w http.ResponseWriter, r *http.Request) {
-	// read all data from reward_delegates table and return
-	edb := srh.GetQueryStateContext().GetEventDB()
-	if edb == nil {
-		common.Respond(w, r, nil, common.NewErrInternal("no db connection"))
-		return
-	}
-
-	challengeID := r.URL.Query().Get("challenge_id")
-	query := r.URL.Query().Get("query")
-
-	var rds []event.RewardDelegate
-
-	if query != "" {
-		rds = edb.RunWhereQueryInDelegateRewards(query)
-	} else if challengeID != "" {
-		rds = edb.GetDelegateChallengeRewardsByID(challengeID)
-	} else {
-		rds = edb.GetAllDelegateChallengeRewards()
-	}
-
-	var sum int64
-
-	for _, rd := range rds {
-		cur, _ := rd.Amount.Int64()
-		sum += cur
-	}
-
-	result := map[string]interface{}{
-		"sum": sum,
-		"rds": rds,
-	}
-
-	// return the list of reward delegates
-	common.Respond(w, r, result, nil)
-}
-
-func (srh *StorageRestHandler) getAllChallenges(w http.ResponseWriter, r *http.Request) {
-	// read all data from challenges table and return
-	edb := srh.GetQueryStateContext().GetEventDB()
-	if edb == nil {
-		common.Respond(w, r, nil, common.NewErrInternal("no db connection"))
-		return
-	}
-
-	allocationID := r.URL.Query().Get("allocation_id")
-
-	challenges, _ := edb.GetAllChallengesByAllocationID(allocationID)
-	common.Respond(w, r, challenges, nil)
-}
-
-func (srh *StorageRestHandler) getBlockRewards(w http.ResponseWriter, r *http.Request) {
-	// read all data from block_rewards table and return
-	edb := srh.GetQueryStateContext().GetEventDB()
-	if edb == nil {
-		common.Respond(w, r, nil, common.NewErrInternal("no db connection"))
-		return
-	}
-
-	blockNumber := r.URL.Query().Get("block_number")
-	startBlockNumber := r.URL.Query().Get("start_block_number")
-	endBlockNumber := r.URL.Query().Get("end_block_number")
-
-	providerRewards := edb.GetBlockRewardsToProviders(blockNumber, startBlockNumber, endBlockNumber)
-	delegateRewards := edb.GetBlockRewardsToDelegates(blockNumber, startBlockNumber, endBlockNumber)
-
-	result := map[string]interface{}{
-		"provider_rewards": providerRewards,
-		"delegate_rewards": delegateRewards,
-	}
-
-	common.Respond(w, r, result, nil)
-}
-
-func (srh *StorageRestHandler) getReadRewards(w http.ResponseWriter, r *http.Request) {
-	// read all data from block_rewards table and return
-	edb := srh.GetQueryStateContext().GetEventDB()
-	if edb == nil {
-		common.Respond(w, r, nil, common.NewErrInternal("no db connection"))
-		return
-	}
-
-	blockNumber := r.URL.Query().Get("block_number")
-	startBlockNumber := r.URL.Query().Get("start_block_number")
-	endBlockNumber := r.URL.Query().Get("end_block_number")
-
-	providerRewards := edb.GetReadRewardsToProviders(blockNumber, startBlockNumber, endBlockNumber)
-	delegateRewards := edb.GetReadRewardsToDelegates(blockNumber, startBlockNumber, endBlockNumber)
-
-	result := map[string]interface{}{
-		"provider_rewards": providerRewards,
-		"delegate_rewards": delegateRewards,
-	}
-
-	common.Respond(w, r, result, nil)
-}
-
-func (srh *StorageRestHandler) getChallengeRewards(w http.ResponseWriter, r *http.Request) {
-	// read all data from challenge_rewards table and return
-	edb := srh.GetQueryStateContext().GetEventDB()
-	if edb == nil {
-		common.Respond(w, r, nil, common.NewErrInternal("no db connection"))
-		return
-	}
-
-	challengeID := r.URL.Query().Get("challenge_id")
-
-	blobberRewards, validatorRewards := edb.GetChallengeRewardsToProviders(challengeID)
-	blobberDelegateRewards, validatorDelegateRewards := edb.GetChallengeRewardsToDelegates(challengeID)
-
-	result := map[string]interface{}{
-		"blobber_rewards":            blobberRewards,
-		"validator_rewards":          validatorRewards,
-		"blobber_delegate_rewards":   blobberDelegateRewards,
-		"validator_delegate_rewards": validatorDelegateRewards,
-	}
-
-	common.Respond(w, r, result, nil)
-}
-
-func (srh *StorageRestHandler) getTotalChallengeRewards(w http.ResponseWriter, r *http.Request) {
-	// read all data from challenge_rewards table and return
-	edb := srh.GetQueryStateContext().GetEventDB()
-	if edb == nil {
-		common.Respond(w, r, nil, common.NewErrInternal("no db connection"))
-		return
-	}
-
-	allocationID := r.URL.Query().Get("allocation_id")
-
-	challenges, _ := edb.GetAllChallengesByAllocationID(allocationID)
-
-	totalBlobberRewards := map[string]int64{}
-	totalValidatorRewards := map[string]int64{}
-
-	for _, challenge := range challenges {
-		blobberRewards, validatorRewards := edb.GetChallengeRewardsToProviders(challenge.ChallengeID)
-		for _, reward := range blobberRewards {
-			// check if the provider_id is already in the map totalBlobberRewards
-			if _, ok := totalBlobberRewards[reward.ProviderId]; ok {
-				cur, _ := reward.Amount.Int64()
-				totalBlobberRewards[reward.ProviderId] += cur
-			} else {
-				cur, _ := reward.Amount.Int64()
-				totalBlobberRewards[reward.ProviderId] = cur
-			}
-		}
-
-		for _, reward := range validatorRewards {
-			// check if the provider_id is already in the map totalChallengeRewards
-			if _, ok := totalValidatorRewards[reward.ProviderId]; ok {
-				cur, _ := reward.Amount.Int64()
-				totalValidatorRewards[reward.ProviderId] += cur
-			} else {
-				cur, _ := reward.Amount.Int64()
-				totalValidatorRewards[reward.ProviderId] = cur
-			}
-		}
-
-		blobberDelegateRewards, validatorDelegateRewards := edb.GetChallengeRewardsToDelegates(challenge.ChallengeID)
-		for _, reward := range blobberDelegateRewards {
-			// check if the provider_id is already in the map totalBlobberRewards
-			if _, ok := totalBlobberRewards[reward.ProviderID]; ok {
-				cur, _ := reward.Amount.Int64()
-				totalBlobberRewards[reward.ProviderID] += cur
-			} else {
-				cur, _ := reward.Amount.Int64()
-				totalBlobberRewards[reward.ProviderID] = cur
-			}
-		}
-
-		for _, reward := range validatorDelegateRewards {
-			// check if the provider_id is already in the map totalChallengeRewards
-			if _, ok := totalValidatorRewards[reward.ProviderID]; ok {
-				cur, _ := reward.Amount.Int64()
-				totalValidatorRewards[reward.ProviderID] += cur
-			} else {
-				cur, _ := reward.Amount.Int64()
-				totalValidatorRewards[reward.ProviderID] = cur
-			}
-		}
-	}
-
-	result := map[string]interface{}{
-		"blobber_rewards":   totalBlobberRewards,
-		"validator_rewards": totalValidatorRewards,
-	}
-
-	common.Respond(w, r, result, nil)
-}
-
-func (srh *StorageRestHandler) getAllocationCancellationReward(w http.ResponseWriter, r *http.Request) {
-	// read all data from allocation_cancellation_rewards table and return
-	edb := srh.GetQueryStateContext().GetEventDB()
-	if edb == nil {
-		common.Respond(w, r, nil, common.NewErrInternal("no db connection"))
-		return
-	}
-
-	startBlock := r.URL.Query().Get("start_block")
-	endBlock := r.URL.Query().Get("end_block")
-
-	providerRewards := edb.GetAllocationCancellationRewardsToProviders(startBlock, endBlock)
-	delegateRewards := edb.GetAllocationCancellationRewardsToDelegates(startBlock, endBlock)
-
-	result := map[string]interface{}{
-		"provider_rewards": providerRewards,
-		"delegate_rewards": delegateRewards,
-	}
-
-	common.Respond(w, r, result, nil)
 }
